requirements.txt
absl-py==1.4.0
anyio==3.6.2
appdirs==1.4.4
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
astunparse==1.6.3
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
bleach==5.0.1
blinker==1.4
Brlapi==0.8.3
Brotli==1.0.9
cachetools==5.3.0
certifi==2020.6.20
cffi==1.15.1
chardet==4.0.0
click==8.0.3
colorama==0.4.4
comm==0.1.2
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
debugpy==1.6.5
decorator==4.4.2
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.6
distro==1.7.0
distro-info===1.1build1
Django==4.1.5
duplicity==0.8.21
entrypoints==0.4
fasteners==0.14.1
fastjsonschema==2.16.2
filelock==3.9.0
flatbuffers==23.1.21
fonttools==4.29.1
fqdn==1.5.1
fs==2.4.12
future==0.18.2
gast==0.4.0
google-auth==2.16.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.51.1
gunicorn==20.1.0
h5py==3.8.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
iniconfig==1.1.1
ipykernel==6.20.1
ipython==7.31.1
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.2
joblib==0.17.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.5
keras==2.11.0
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==15.0.6.1
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.4
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.7
nbformat==5.7.3
nest-asyncio==1.5.6
netifaces==0.11.0
networkx==3.0
nltk==3.8.1
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
opencv-python==4.7.0.68
opt-einsum==3.3.0
packaging==21.3
pandas==1.5.2
pandocfilters==1.5.0
paramiko==2.9.3
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.6.2
pluggy==0.13.0
ply==3.11
prometheus-client==0.15.0
prompt-toolkit==3.0.28
protobuf==3.12.4
psutil==5.9.4
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.20.1
pycparser==2.21
pycups==2.0.1
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.19.3
pytest==6.2.5
python-apt==2.4.0
python-dateutil==2.8.2
python-debian===0.1.43ubuntu1
python-json-logger==2.0.4
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==25.0.0
qtconsole==5.4.0
QtPy==2.3.0
regex==2022.10.31
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==0.23.2
scipy==1.8.0
SecretStorage==3.3.1
Send2Trash==1.8.0
simplejson==3.17.6
six==1.16.0
sklearn==0.0.post1
sniffio==1.3.0
soupsieve==2.3.1
sqlparse==0.4.3
summary==0.2.0
sympy==1.9
systemd-python==234
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.30.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
ubuntu-advantage-tools==27.12
ubuntu-drivers-common==0.0.0
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
uri-template==1.2.0
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.17.1
wadllib==1.3.6
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.5
wrapt==1.14.1
xdg==5
xgboost==1.7.3
xkit==0.0.0
zipp==1.0.0
